add new karan rul (wip)
