+++ src/tokenizer.rs
@@ -1,5 +1,6 @@
 mod dict_reader;
 pub mod newmm;
 mod tcc;
+pub(crate) mod tcc_rules;
 pub mod tokenizer_trait;
 mod trie_char;
