+++ src/tokenizer/tcc.rs
@@ -10,60 +10,15 @@ Credits:
     * Python code: Korakot Chaovavanich
     * Rust Code Translation: Thanathip Suntorntip
 */
-use lazy_static::lazy_static;
-use regex::bytes::Regex;
+
+
 use rustc_hash::FxHashSet as HashSet;
 
-use crate::four_bytes_str::custom_regex::{regex_pattern_to_custom_pattern, replace_tcc_symbol};
+use crate::four_bytes_str::custom_regex::{regex_pattern_to_custom_pattern};
 use crate::four_bytes_str::custom_string::{
     CustomStringBytesSlice, FixedCharsLengthByteSlice, BYTES_PER_CHAR,
 };
-
-// regex crate does not support look-any-direction
-
-lazy_static! {
-    static ref NON_LOOKAHEAD_TCC: Regex = Regex::new(
-        &[
-            r"^เc็c", //1
-            r"^เcctาะ",//2
-            r"^เccีtยะ",//3
-            r"^เcc็c",//4
-            r"^เcิc์c",//5
-            r"^เcิtc",//6
-            r"^เcีtยะ?",//7
-            r"^เcืtอะ?",//8
-            r"^เctา?ะ?",//9
-            r"^cัtวะ",//10
-            r"^c[ัื]tc[ุิะ]?",//11
-            r"^c[ิุู]์",//12
-            r"^c[ะ-ู]t",//13
-            r"^c็",//14
-            r"^ct[ะาำ]?",//15
-            r"^แc็c",//16
-            r"^แcc์",//17
-            r"^แctะ",//18
-            r"^แcc็c",//19
-            r"^แccc์",//20
-            r"^โctะ",//21
-            r"^[เ-ไ]ct",//22
-            r"^(เccีtย)[เ-ไก-ฮ]", // look ahead 1
-            r"^(เc[ิีุู]tย)[เ-ไก-ฮ]",// look ahead 2
-            ].map(|pattern| {regex_pattern_to_custom_pattern(&replace_tcc_symbol(pattern)).unwrap()})
-        .join("|")
-    )
-    .unwrap();
-}
-
-lazy_static! {
-    static ref LOOKAHEAD_TCC: Regex = Regex::new(
-        &[
-            r"^(เccีtย)[เ-ไก-ฮ]",//เccีtย(?=[เ-ไก-ฮ]|$)
-            r"^(เc[ิีุู]tย)[เ-ไก-ฮ]",//เc[ิีุู]tย(?=[เ-ไก-ฮ]|$)
-        ].map(|pattern| {regex_pattern_to_custom_pattern(&replace_tcc_symbol(pattern)).unwrap()})
-        .join("|")
-    )
-    .unwrap();
-}
+use super::tcc_rules::{L,NON_LOOKAHEAD_TCC};
 
 pub fn tcc_pos(custom_text_type: &CustomStringBytesSlice) -> HashSet<usize> {
     let mut set: HashSet<usize> = HashSet::default();
@@ -100,3 +55,7 @@ pub fn tcc_pos(custom_text_type: &CustomStringBytesSlice) -> HashSet<usize> {
     }
     set
 }
+
+// pub fn tcc_pos_p(custom_text_type: &CustomStringBytesSlice) -> HashSet<usize> {
+
+// }
